requirements.txt
Django==1.8.7
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-filter==0.11.0
django-rest-swagger==0.3.4
djangorestframework==3.3.1
djangorestframework-filters==0.6.0
gunicorn==19.4.1
MySQL-python==1.2.5
netaddr==0.7.18
pydot==1.0.2
pyparsing==2.0.6
PyYAML==3.11
six==1.10.0
sqlparse==0.1.18
wheel==0.24.0
